+++ CHANGELOG.md
@@ -19,6 +19,11 @@ changes.
 
 - Tested with `cardano-node 10.5.3` and `cardano-cli 10.11.0.0`.
 
+- **BREAKING** Hydra node stops interacting on network and client API once the node's view of the chain is out of sync more than 50% of `--contestation-period`.
+  - Added `NodeUnsynced` and `NodeSynced` state events and server outputs.
+  - Added `RejectedInput` client message.
+  - The `Checkpoint` event, and consequently the `EventLogRotated` server output, now carry the different `NodeState` variants: `NodeInSync` or `NodeCatchingUp`.
+  - `Greetings` message now also contains the hydra-node synced status to the chain backend.
 
 ## [1.1.0] - 2025-10-28